// apps/mobile/lib/api/model/internalHandlerHttpNotificationPreferencesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Wakeup API
 * Friend-graph chat backend. See docs/WAKEUP.md for the full spec.
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerHttpNotificationPreferencesResponseThemeModePreference } from './internalHandlerHttpNotificationPreferencesResponseThemeModePreference';
import type { InternalHandlerHttpNotificationPreferencesResponseThemeScheme } from './internalHandlerHttpNotificationPreferencesResponseThemeScheme';

export interface InternalHandlerHttpNotificationPreferencesResponse {
  calls?: boolean;
  direct_messages?: boolean;
  friend_requests?: boolean;
  group_messages?: boolean;
  theme_mode_preference?: InternalHandlerHttpNotificationPreferencesResponseThemeModePreference;
  theme_scheme?: InternalHandlerHttpNotificationPreferencesResponseThemeScheme;
}